afterscan page is done, aside from the upper limmit and RI text explanations. For some reason the app crashes, when using those pictures.
the layout is called activity_database.xml and the coding file is called DatabaseActivity.kt.

In the FoodDao.kt a new Query has been added to use for the test, as we are picking specific food items. it looks through the database based on the name of the food product.

on the startup page i added a button that can be used, if no bottle is available or if you are testing in the emulator. It skips the scanning part and goes directly to the Vitamin C information page.